+++ test/routes/index.test.js
@@ -551,6 +551,7 @@ describe('getRouteHandlers', () => {
     listHistory: sinon.stub(),
     createMapping: sinon.stub(),
     patchMapping: sinon.stub(),
+    deleteMapping: sinon.stub(),
     getProductCapabilities: sinon.stub(),
     getUserCapabilities: sinon.stub(),
     getAuditLogs: sinon.stub(),
@@ -821,6 +822,7 @@ describe('getRouteHandlers', () => {
 
     const expectedDynamicRouteKeys = [
       'PATCH /state/access-mappings/:id',
+      'DELETE /state/access-mappings/:id',
       'GET /user/capabilities/:resourceId',
       'GET /organizations/:organizationId/permission/audit-logs',
       'GET /audits/latest/:auditType',